Modify git prompt to be more customizable.
Refactor the git prompt script in order to logically split the git status engine
from the git prompt display. Use associative arrays to store the git status
variables, the symbols and colors to be used, as well as the order in which the
symbols appear. Additionally, add functionality to depict numbers as superscript
or subscript. Finally, add functionality to choose how a false flag information
will be depicted.